requirements.txt
altgraph==0.17.4
contourpy==1.2.1
cycler==0.12.1
fonttools==4.53.1
kiwisolver==1.4.5
matplotlib==3.9.2
more-itertools==10.4.0
numpy==2.0.1
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pygame-ce==2.5.1
pygame_gui==0.6.12
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-i18n==0.3.9
pytz==2024.1
seaborn==0.13.2
setuptools==75.1.0
six==1.16.0
tzdata==2024.1
